+++ src/include/coati/mutation_ecm.hpp
@@ -26,8 +26,8 @@
 #include "mutation_fst.hpp"
 
 void nts_ntv(uint8_t c1, uint8_t c2, int& nts, int& ntv);
-double k(uint8_t c1, uint8_t c2, int model = 0);
-void ecm_p(Matrix64f& P, double br_len);
-void ecm(VectorFstStdArc& mut_fst, double br_len);
+float k(uint8_t c1, uint8_t c2, int model = 0);
+void ecm_p(Matrix64f& P, float br_len);
+void ecm(VectorFstStdArc& mut_fst, float br_len);
 
 #endif
